[BE] refactor: diary pacade 패턴 적용
- 단일 책임 원칙을 위해 서비스 간 분리하는 과정 중 발생한 순환 참조로 인하여 pacade 디자인 패턴 적용
